Add git-config script
Groovy version required: 3.X.X
This script have the purpose to of change between configurations for
local repositories by creating a configuration in JSON format with the
different profiles.

This is still a work in progress, some functionalities may not work as
expected.

Warning: This was tested only on Mac env, not tested yet for other OS.